+++ Laravel/app/Http/Requests/UpdateTeamRequest.php
@@ -30,8 +30,12 @@ class UpdateTeamRequest extends FormRequest
     public function rules()
     {
         return [
-            'name' => ['sometimes', 'required', 'string', 'max:25'],
-            'position' => ['sometimes', 'required', 'string', 'max:25'],
+            'en_name' => ['sometimes', 'required', 'string', 'max:25'],
+            'ku_name' => ['sometimes', 'required', 'string', 'max:25'],
+            'ar_name' => ['sometimes', 'required', 'string', 'max:25'],
+            'en_position' => ['sometimes', 'required', 'string', 'max:25'],
+            'ku_position' => ['sometimes', 'required', 'string', 'max:25'],
+            'ar_position' => ['sometimes', 'required', 'string', 'max:25'],
             'image' => [
                 'sometimes', 'image', 'max:2048',
                 'mimes:jpeg,jpg,png,svg',
@@ -39,6 +43,22 @@ class UpdateTeamRequest extends FormRequest
             'facebook' => ['sometimes', 'nullable', 'string', 'url'],
             'instagram' => ['sometimes', 'nullable', 'string', 'url'],
             'twitter' => ['sometimes', 'nullable', 'string', 'url'],
+        ];
+    }
+    /**
+     * Get custom attributes for validator errors.
+     *
+     * @return array
+     */
+    public function attributes()
+    {
+        return [
+            'en_name' => "English Name",
+            'ku_name' => "Kurdish Name",
+            'ar_name' => "Arabic Name",
+            'en_position' => "English Position",
+            'ku_position' => "Kurdish Position",
+            'ar_position' => "Arabic Position",
         ];
     }
     public function updateRecord(Team $team)
@@ -50,22 +70,23 @@ class UpdateTeamRequest extends FormRequest
             $inputs->put('image', $image);
         }
         $member = $team->update($inputs->only([
-            'name', 'position', 'socials', 'image'
+            'en_name', 'en_position', 'ku_name', 'ku_position', 'ar_name', 'ar_position', 'socials', 'image'
         ])->toArray());
 
         if (!$member)
             throw ValidationException::withMessages([
-                'failed' => "Failed to update the team member"
+                'failed' => __('index.admin.messages.team.fail.update')
             ]);
     }
 
     function uploadImage(string $path)
     {
-        if (file_exists($path)) {
+        $notDefault = \preg_match('/(default.png)/', $path);
+        if (file_exists($path) && ($notDefault === 0)) {
             $real_path = str_replace('uploads/', '', $path);
             Storage::disk('public')->delete($real_path);
         }
-        return 'uploads/' . $this->file('image')->store('images/team', 'public');
+        return 'uploads/' . $this->file('image')->store('images/teams', 'public');
     }
     function mergeSocials()
     {
